adjust ui - composition
